Made use of InputVariable object in both htc.input and htc.inputfile.
