+++ abipy/htc/input.py
@@ -19,6 +19,7 @@ from abipy.core import Structure
 from abipy.tools import is_string, list_strings
 from pymatgen.core.units import Energy
 
+from .variable import InputVariable
 
 __all__ = [
     "AbiInput",
@@ -655,7 +656,8 @@ class Dataset(collections.Mapping):
                 continue
 
             varname = var + post
-            app("%s %s" % (varname, format_var(var, value)))
+            variable = InputVariable(varname, value)
+            app(str(variable))
 
         return "\n".join(lines)
 
@@ -818,316 +820,6 @@ class Dataset(collections.Mapping):
                            nkptgw=nkptgw,
                            bdgw=np.reshape(bdgw, (nkptgw, 2)),
                            )
-
-
-def format_var(varname, value):
-    """
-    Return a string for a single input variable to be written in the input file.
-    Always ends with a carriage return.
-    """
-    if value is None or not str(value):
-        return ""
-
-    # By default, do not impose a number of decimal points ...but not for those
-    digits = {
-        ('xred','xcart','qpt','kpt') : 10,
-        ('ngkpt','kptrlatt', 'ngqpt'): 0,
-    }
-
-    precision = {}
-    for t, numd in digits.items():
-        for varname in t:
-            precision[varname] = numd 
-
-    floatdecimal = precision.get(varname, 0)
-
-    line = " "
-
-    if isinstance(value, np.ndarray):
-        n = 1
-        for i in np.shape(value):
-            n *= i
-        value = list(np.reshape(value, n))
-
-    if isinstance(value, (list, tuple)):
-        # values in lists
-        # Reshape a list of lists into a single list
-        if all(isinstance(v, list) for v in value):
-            line += format_list_of_list(value, floatdecimal)
-
-        else:
-            # Maximum number of values per line.
-            valperline = 3
-            if any(inp in varname for inp in ['bdgw']):
-                valperline = 2
-
-            line += format_list(value, valperline, floatdecimal)
-
-    else:
-        # scalar values
-        line += ' ' + str(value)
-
-    return line
-
-
-def format_scalar(val, floatdecimal=0):
-    """
-    Format a single numerical value into a string
-    with the appropriate number of decimals.
-    """
-    sval = str(val)
-    if sval.isdigit() and floatdecimal == 0:
-        return sval
-
-    try:
-        fval = float(val)
-    except:
-        return sval
-
-    if fval == 0 or (abs(fval) > 1e-3 and abs(fval) < 1e4):
-        form = 'f'
-        addlen = 5
-    else:
-        form = 'e'
-        addlen = 8
-
-    ndec = max(len(str(fval-int(fval)))-2, floatdecimal)
-    ndec = min(ndec, 10)
-
-    sval = '{v:>{l}.{p}{f}}'.format(v=fval, l=ndec+addlen, p=ndec, f=form)
-
-    return sval.replace('e', 'd')
-
-
-def format_list_of_list(values, floatdecimal=0):
-    """Format a list of lists."""
-
-    lvals = flattens(values)
-
-    # Determine the representation
-    if all(isinstance(v, int) for v in lvals):
-        type_all = int
-    else:
-        try:
-            for v in lvals:
-                float(v)
-            type_all = float
-        except:
-            type_all = str
-
-    # Determine the format
-    width = max(len(str(s)) for s in lvals)
-    if type_all == int:
-        formatspec = '>{}d'.format(width)
-
-    elif type_all == str:
-        formatspec = '>{}'.format(width)
-
-    else:
-        # Number of decimal
-        maxdec = max(len(str(f-int(f)))-2 for f in lvals)
-        ndec = min(max(maxdec, floatdecimal), 10)
-
-        if all(f == 0 or (abs(f) > 1e-3 and abs(f) < 1e4) for f in lvals):
-            formatspec = '>{w}.{p}f'.format(w=ndec+5, p=ndec)
-        else:
-            formatspec = '>{w}.{p}e'.format(w=ndec+8, p=ndec)
-
-    line = '\n'
-    for L in values:
-        for val in L:
-            line += ' {v:{f}}'.format(v=val, f=formatspec)
-        line += '\n'
-
-    return line
-
-
-def format_list(values, valperline=3, floatdecimal=0):
-    """
-    Format a list of values into a string.  The result might be spread among
-    several lines, each line starting with a blank space.
-    'valperline' specifies the number of values per line.
-    """
-    line = ""
-    # Format the line declaring the value
-    for i, val in enumerate(values):
-        line += ' ' + format_scalar(val, floatdecimal)
-        if (i+1) % valperline == 0:
-            line += '\n'
-
-    # Add a carriage return in case of several lines
-    if '\n' in line.rstrip('\n'):
-        line = '\n' + line
-
-    # Add a carriage return at the end
-    if not line.endswith('\n'):
-        line += '\n'
-
-    return line
-
-
-def string_to_value(sval):
-    """
-    Interpret a string variable and attempt to return a value of the appropriate type.  
-    If all else fails, return the initial string.
-    """
-    value = None
-    try:
-        for part in sval.split():
-            if '*' in part:
-                if part[0] == '*':
-                    # cases like istwfk *1
-                    value = None
-                    break
-
-                else:
-                    # cases like acell 3*3.52
-                    n = int(part.split('*')[0])
-                    f = convert_number(part.split('*')[1])
-                    if value is None:
-                        value = []
-                    value += n * [f]
-                    continue
-
-            if '/' in part:
-                # Fractions
-                (num, den) = (float(part.split('/')[i]) for i in range(2))
-                part = num / den
-
-            if part in _UNITS.keys():
-                # Unit
-                if value is None:
-                    msg = "Could not apply the unit tolken '%s'." %(part)
-                    warnings.warn(msg)
-                elif isinstance(value, list):
-                    value.append(part)
-                else:
-                    value = [value, part]
-
-                # Convert
-                if False:
-                    if isinstance(value, (list, tuple)):
-                        for i in range(len(value)):
-                            value[i] *= _UNITS[part]
-                    elif is_string(value):
-                        value = None
-                        break
-                    else:
-                        value *= _UNITS[part]
-
-                continue
-
-            # Convert
-            try:
-                val = convert_number(part)
-            except:
-                val = part
-
-            if value is None:
-                value = val
-            elif isinstance(value, list):
-                value.append(val)
-            else:
-                value = [value, val]
-    except:
-        value = None
-
-    if value is None:
-        value = sval
-
-    return value
-
-
-def format_variable_dict(variables):
-    """Format a dictionary of input variables into a multi-line string."""
-    s = ""
-    for (var, val) in variables.items():
-        s += format_var(var, val)
-
-    return s
-
-
-def flattens(lists):
-    """Transform a list of lists into a single list."""
-    if not isinstance(lists[0], list):
-        return lists
-    lst = []
-    for l in lists:
-        lst += l
-
-    return lst
-
-
-def listify(obj):
-    """Transform any object, iterable or not, to a list."""
-    if isinstance(obj, collections.Iterable):
-        return obj
-
-    elif isinstance(obj, collections.Iterator):
-        return list(obj)
-
-    else:
-        return [obj]
-
-#def listify(obj):
-#    """
-#    Transform any object, iterable or not, to a list. 
-#    Returns obj if object is Iterable with the exception of 
-#    single string that is converted to list.
-#    """
-#    if is_string(obj)
-#       return [obj]
-#
-#    if isinstance(obj, collections.Iterable):
-#        if isinstance(obj, tuple):
-#            return list(obj)
-#        else:
-#            return obj
-#
-#    elif isinstance(obj, collections.Iterator):
-#        return list(obj)
-#
-#    else:
-#        return [obj]
-
-
-def is_number(s):
-    """Returns True if the argument can be converted to float."""
-    try:
-        float(s)
-        return True
-    except:
-        return False
-
-
-def convert_number(value):
-    """
-    Converts some object to a float or a string.
-    If the argument is an integer or a float, returns the same object.
-    If the argument is a string, tries to convert to an integer, then to a float.
-    The string '1.0d-03' will be treated the same as '1.0e-03'
-    """
-    if isinstance(value (float, int)):
-        return value
-
-    elif is_string(value):
-
-        if is_number(value):
-            try:
-                return int(value)
-            except ValueError:
-                return float(value)
-
-        else:
-            val = value.replace('d', 'e')
-            if is_number(val):
-                return float(val)
-            else:
-                raise ValueError("convert_number failed")
-
-    else:
-        raise ValueError("convert_number failed")
-
 
 
 class LujForSpecie(collections.namedtuple("LdauForSpecie", "l u j unit")):
